requirements.txt
# Packages for working with YouTube Data API
google-api-python-client==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6

# Packages for making HTTP requests to Reddit API
requests==2.26.0

# MongoDB driver for Python
pymongo==3.12.0

# Other necessary packages
python-dotenv==1.0

# Data visualization for future not for this project.
numpy==1.26.1
pandas==2.1.2
matplotlib==3.4.3


pytz==2023.3
